+++ branches/version_2_double/DrawEngine.Renderer/Mathematics/Algebra/BarycentricCoordinate.cs
@@ -5,11 +5,11 @@ namespace DrawEngine.Renderer.Algebra
     [Serializable, StructLayout(LayoutKind.Sequential, Pack = 1)]
     public struct BarycentricCoordinate
     {
-        public float Alpha;
-        public float Beta;
-        public float Gama;
+        public double Alpha;
+        public double Beta;
+        public double Gama;
         public static readonly BarycentricCoordinate Zero = new BarycentricCoordinate();
-        public BarycentricCoordinate(float alpha, float beta, float gama)
+        public BarycentricCoordinate(double alpha, double beta, double gama)
         {
             this.Alpha = alpha;
             this.Beta = beta;
